Some changes and fixed some bugs

+++ RealEstateApp.Core.Application/Features/TypeOfSales/Command/DeleteTypeOfSaleById/DeleteTypeOfSaleByIdCommand.cs
@@ -1,6 +1,8 @@
 ﻿using MediatR;
+using RealEstateApp.Core.Application.Exceptions;
 using RealEstateApp.Core.Application.Interfaces.Repositories;
 using RealEstateApp.Core.Application.Wrappers;
+using System.Net;
 
 namespace RealEstateApp.Core.Application.Features.TypeOfSales.Command.DeleteTypeOfSaleById
 {
@@ -19,6 +21,7 @@ namespace RealEstateApp.Core.Application.Features.TypeOfSales.Command.DeleteType
         public async Task<Response<Unit>> Handle(DeleteTypeOfSaleByIdCommand command, CancellationToken cancellationToken)
         {
             var typeOfSale = await _typeOfSaleRepository.GetByIdAsync(command.Id);
+            if (typeOfSale is null) throw new ApiException("Type of sale not found", (int)HttpStatusCode.NoContent);
             await _typeOfSaleRepository.DeleteAsync(typeOfSale);
             return new Response<Unit>(Unit.Value);
         }
